ServerlessRAG_BedrockAgents/fastapi_backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
requests>=2.31.0
email-validator>=2.1.0
